bot/requirements.txt
# FeralLobster Telegram Bot 依赖
# Python 3.11+

# Telegram Bot 框架
python-telegram-bot[ext,job-queue]>=20.0

# 加密库
cryptography>=41.0.0

# HTTP 客户端
aiohttp>=3.9.0

# 环境变量管理
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# 工具库
uuid>=1.30

# 可选: 性能优化
uvloop>=0.19.0; sys_platform != "win32"

# 可选: 错误追踪
sentry-sdk>=1.39.0
